Adding packages to be installed
This is adding the following packages to be installed:

libuuid-perl linux-base initramfs-tools irqbalance-imx

The Debian created linux images depend on libuuid-perl, linux-base, and initramfs-tools. Based on that, I think it is a good idea to also have them for the installer.

I also added irqbalance-imx, as I know this is in the Novena repo and is supposed to help improve performance for the Novena.